Add documentation to plane shape families and simplify API.
